requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
PyJWT==2.8.0
Werkzeug==3.0.1
APScheduler==3.10.4
openai==1.14.2
anthropic==0.21.3
httpx==0.27.0
Pillow==10.2.0
requests==2.31.0

# 本地AI生成（以下为可选的本地/推理模型支持，如仅运行基础论坛可暂时跳过安装）
diffusers==0.27.2
transformers==4.38.2
accelerate==0.27.2
torch==2.2.1
gTTS==2.5.0
pydub==0.25.1
numpy==1.26.4
scipy==1.13.0
